+++ mne/viz/tests/test_3d.py
@@ -53,8 +53,7 @@ warnings.simplefilter('always')  # enable b/c these tests throw warnings
 @requires_pysurfer
 @requires_mayavi
 def test_plot_sparse_source_estimates():
-    """Test plotting of (sparse) source estimates
-    """
+    """Test plotting of (sparse) source estimates."""
     sample_src = read_source_spaces(src_fname)
 
     # dense version
@@ -91,8 +90,7 @@ def test_plot_sparse_source_estimates():
 @testing.requires_testing_data
 @requires_mayavi
 def test_plot_evoked_field():
-    """Test plotting evoked field
-    """
+    """Test plotting evoked field."""
     evoked = read_evokeds(evoked_fname, condition='Left Auditory',
                           baseline=(-0.2, 0.0))
     evoked = pick_channels_evoked(evoked, evoked.ch_names[::10])  # speed
@@ -107,8 +105,7 @@ def test_plot_evoked_field():
 @testing.requires_testing_data
 @requires_mayavi
 def test_plot_trans():
-    """Test plotting of -trans.fif files and MEG sensor layouts
-    """
+    """Test plotting of -trans.fif files and MEG sensor layouts."""
     evoked = read_evokeds(evoked_fname)[0]
     with warnings.catch_warnings(record=True):  # 4D weight tables
         bti = read_raw_bti(pdf_fname, config_fname, hs_fname, convert=True,
@@ -136,7 +133,9 @@ def test_plot_trans():
     # EEG only with strange options
     with warnings.catch_warnings(record=True) as w:
         plot_trans(evoked.copy().pick_types(meg=False, eeg=True).info,
-                   trans=trans_fname, meg_sensors=True)
+                   subject='sample', trans=trans_fname, meg_sensors=True,
+                   eeg_sensors=['original', 'projected'],
+                   subjects_dir=subjects_dir)
     assert_true(['Cannot plot MEG' in str(ww.message) for ww in w])
 
 
@@ -144,8 +143,7 @@ def test_plot_trans():
 @requires_pysurfer
 @requires_mayavi
 def test_limits_to_control_points():
-    """Test functionality for determing control points
-    """
+    """Test functionality for determing control points."""
     sample_src = read_source_spaces(src_fname)
 
     vertices = [s['vertno'] for s in sample_src]
@@ -215,8 +213,7 @@ def test_limits_to_control_points():
 @testing.requires_testing_data
 @requires_mayavi
 def test_plot_dipole_locations():
-    """Test plotting dipole locations
-    """
+    """Test plotting dipole locations."""
     dipoles = read_dipole(dip_fname)
     trans = read_trans(trans_fname)
     dipoles.plot_locations(trans, 'sample', subjects_dir, fig_name='foo')